Extract data fetching into useFetchData hook
